requirements.txt
Faker==1.0.7
mysql-connector-python==8.0.16
protobuf==3.8.0
python-dateutil==2.8.0
six==1.12.0
text-unidecode==1.2
